renomeia funcao

+++ modelagemZika.h
@@ -20,6 +20,6 @@
     bool atendeCriterioParadaY(const double resistenciaVascular[], const int rota[], int arteriaAtual, double posYAtual);
     bool atendeCriterioParadaX(const double totalD[], const double comprimentoArteria[], const int rota[], int arteriaAtual, double posXAtual);
     bool atendeCriterioAbsorcaoReflexao(const double resistenciaVascular[], const int rota[], const int arteriaAtual, double posAtualY);
-    std::string numeroPorExtenso(long long numero);
+    std::string numeroPorExtenso(int numero);
 
 #endif
